requirements.txt
fastmcp>=1.0.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
authlib>=1.3.0
tenacity>=8.2.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
aiofiles>=23.0.0
